AWS生成AIアプリ構築実践ガイド/src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.19",
    "boto3-stubs>=1.42.19",
    "datasets==3.0.2",
    "langchain==0.3.21",
    "langchain-aws==0.2.18",
    "langchain-text-splitters>=0.3.11",
    "mypy-boto3-bedrock-agent>=1.42.3",
    "mypy-boto3-bedrock-agent-runtime>=1.42.3",
    "mypy-boto3-bedrock-runtime>=1.42.3",
    "pytest>=9.0.2",
    "python-dotenv==1.0.1",
    "ragas==0.2.14",
    "streamlit==1.39.0",
]

[dependency-groups]
dev = [
    "pytest-dotenv>=0.5.2",
]

[tool.pytest.ini_options]
addopts = "-s"
env_files = [".env"]
pythonpath = "."